@@ -9,7 +9,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "h5netcdf"
 description = "netCDF4 via h5py"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = {file = "LICENSE.txt"}
 authors = [
     { name = "Stephan Hoyer", email = "shoyer@gmail.com" },
@@ -25,7 +25,6 @@ classifiers=[
     "Intended Audience :: Science/Research",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
